Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int n1,n2;
+    float n1,n2;
     scanf("%d %d", &n1, &n2);
 
     if((a/b) == b){
